Scopes: Add inline and block scopes
Summary:
Restructures how regex matching works to support block and inline scopes.
This is the start of some work that will allow us to remove hacks from
parsing multi-defs and lists, and should make writing some of these
regexes easier.

To accomplish this, we replaced the `regex:` field with a `match:` field,
which is a function rather than a regex, which takes both the content
and the state object. We then store an `inline` field on the state object,
so that the match function can check whether we are inline or block.

Then we add several helper functions to make the syntax of using these
inline vs block parses easier.

Test Plan:
`make test` both here and in perseus

Opened perseus and typed some things in the box--adding various
syntax elements and a widget. Didn't see anything obviously
broken.

Reviewers: emily, alex

Reviewed By: alex

Subscribers: emily, alex

Differential Revision: https://phabricator.khanacademy.org/D15959