Priorities: Rework priorities for more objects; less splicing
Summary:
Changes the way priorities work. First, they are now
called `order`. `order` is now a number on each rule object
which allows for sorting. `SimpleMarkdown.parserFor()`
creates the rule order list based on these `order` values.

For extensions, extenders should add a `order:` field
to their rules, which should probably be populated with
a number relative to a preexisting rule order, such as

    SimpleMarkdown.defaultRules.em.order - 0.5

I renamed these to `order` because having a lower priority
mean that a rule was processed earlier felt confusing, while
having the number be an ascending number seemed to make much
more sense code-wise. I'm not sure order is the clearest
possible name, but it seemed to alleviate this problem.

Test Plan:
`make test`

I added a new extension test, as well as a few test for sorting
the orders.

Reviewers: alex

Reviewed By: alex

Subscribers: emily

Differential Revision: https://phabricator.khanacademy.org/D15988