[react] Use inline ReactElement literals
Summary:
This commit inlines the results of React.createElement! This is a tad
sketchy, but has a couple useful properties.

Mainly, it allows us to remove the dependency on React, which is useful for:

 * when you're using html output (to be added soon)
 * you have your own way of requiring React, that might
   be nonstandard (window.React? require("react")?)

This also makes output a little faster, which, given how large markdown output
can be, is probably nice.

Clients should be able to use normal jsx/React.createElement syntax.

Depends on D16647

Test Plan:
    make test

Reviewers: alex

Reviewed By: alex

Differential Revision: https://phabricator.khanacademy.org/D16649