Add ability to output plain HTML from markdown
Summary:
As the title mentions, now we can output plain HTML using
simple-markdown!

This was done by adding an `html` function to each of the rules that
needed it and then an `htmlFor` function that behaves similarly to the
`reactFor` function.

The `SimpleMarkdown` object now has two additional public methods:
`htmlFor` and `defaultHtmlOutput` which are analogous to `reactFor` and
`defaultReactOutput`.

I also copied over the tests for React output and changed 'em so they
correctly test HTML output as well.

Test Plan:
Verify that `make test` passes.

Reviewers: aria