Fix backslash in `[^` block
This backslash was being ignored because `\_` is the same as `_`.
It should have been `\\_` in the regex and `"\\\\_"` in the string,
so that it only accepts backslashes followed by characters, and the
`[^` block never tries to take a backslash and then has to backtrack
for the `\\\\[\\s\\S]` block :).

Also bumped the version for this bugfix