Output: change how output arrays are canonically joined
This commit introduces a new way to specify how output
is to be joined, using `outputFor` and the `Array` rule,
instead of `reactFor` and `ruleOutput`.

This approach adds more customization ability of how nodes
in an array are joined (or not) for output.

This change enables us to modify how text nodes are joined
by default to improve react-native support, since in
react-native we use <Text> elements instead of raw strings.

This shouldn't affect previous usage, but it is encouraged
to use the new `outputFor` interface when building a
custom output function.