parserFor/outputFor: add defaultState param
Adds an optional `defaultState` parameter to `parserFor` and
`outputFor`.

This simplifies common usage patterns without requiring clients
to create their own wrapper functions.

Note that this has to mutate the passed in state so that same
object is preserved for examining after the parse/output.
This is a little inelegant, but powerful (and used in
several existing extensions)