Strikethrough: fix exponential backtracking
A long sequence of backslashes inside a strikethrough could confuse the
strikethrough regex into exponential backtracking, causing a potential
ReDoS vulnerability.

This commit updates the strikethrough regex to only accept a backslash
if it is preceding an escaped character, as other rules handle
backslashes.

Updates to version 0.7.3 to publish this fix.

Thanks to @pwntester and the [GitHub Security Lab team](https://securitylab.github.com/)
for finding this vulnerability!

Test plan:

1. `make test`
    * verify the new strikethrough backtracking test passes
    * verify all the prior tests pass